requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.5.0
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
responses>=0.25.0